backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
yt-dlp>=2024.1.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
